refactor(pa-71): make signUp and signIn pages look different
