+++ frontend/src/components/authentication/SignUp.styles.tsx
@@ -3,7 +3,8 @@ import { makeStyles } from '@material-ui/core/styles'
 
 export const useStyles = makeStyles((theme) => ({
     root: {
-      height: '100vh',
+      display: 'flex',
+      justifyContent: 'space-around',
     },
     image: {
       backgroundImage: 'url(https://source.unsplash.com/random)',
@@ -31,3 +32,5 @@ export const useStyles = makeStyles((theme) => ({
       margin: theme.spacing(3, 0, 2),
     },
   }))
+
+
